Delete some redundant `activemq` native-image conf
These deletions were performed by trial-and-error. It may be possible to
further reduce this configuration, but I'm going to consider this
sufficient for now.

I found most of the configuration to be redundant, so I removed it. What
remains is: a pair of proxies and reflections from within the following
four packages:

- `com.example.activemq`
- `javax.jms`
- `org.apache.activemq`
- `org.springframework.jms`